slangmang/requirements.txt
appnope==0.1.0
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
colorama==0.4.3
coverage==5.0.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.4
django-cors-headers==3.2.1
django-phone-field==1.8.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==4.0.2
green==3.1.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
Unidecode==1.1.1
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.1.0
